tests/wpt/web-platform-tests/svg/painting/parsing/fill-rule-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: getComputedStyle().fillRule</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#FillRuleProperty"/>
    <h:meta name="assert" content="fill-rule computed value is as specified."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("fill-rule", "nonzero");
test_computed_value("fill-rule", "evenodd");

  ]]></script>
</svg>
